tecsus/requirements.txt
argcomplete==3.3.0
asgiref==3.8.1
charset-normalizer==3.3.2
colorama==0.4.6
commitizen==3.27.0
decli==0.6.2
distlib==0.3.8
Django==5.0.6
django-cors-headers==4.3.1
djangorestframework==3.15.1
filelock==3.13.1
importlib_metadata==7.1.0
iniconfig==2.0.0
Jinja2==3.1.4
MarkupSafe==2.1.5
packaging==24.0
platformdirs==4.1.0
pluggy==1.5.0
prompt-toolkit==3.0.36
psycopg2==2.9.9
pyasn1==0.6.0
pytest==8.2.0
pytest-django==4.8.0
python-decouple==3.8
PyYAML==6.0.1
questionary==2.0.1
rsa==4.9
sqlparse==0.5.0
termcolor==2.4.0
tomlkit==0.12.5
tzdata==2024.1
virtualenv==20.25.0
wcwidth==0.2.13
zipp==3.18.2
